refactor: Repository interface 로 분리
